kood/chatbot-workshop/pyproject.toml
[tool.poetry]
name = "chatbot-workshop"
version = "0.1.0"
description = ""
authors = ["mathiasare <mathias.are@ridango.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
discord = "^2.2.2"
python-dotenv = "^1.0.0"
requests = "^2.28.2"
openai = "^0.27.4"
estnltk = "^1.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
